requirements.txt
# Web框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# 大模型编排
langchain>=0.2.0
langgraph>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# 数据库与存储
GitPython>=3.1.0

# 机器学习
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
tensorflow>=2.13.0
keras>=2.13.0
shap>=0.42.0

# 工具链
joblib>=1.3.0
tiktoken>=0.5.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
python-docx>=0.8.0

# 前端
streamlit>=1.28.0
plotly>=5.17.0

# 配置与日志
pyyaml>=6.0.1
python-json-logger>=2.0.7

# 向量检索与RAG
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# 网络爬虫
requests>=2.31.0
beautifulsoup4>=4.12.0

# NLP与深度学习
torch>=2.0.0
pytorch-crf>=0.7.0

# 可视化
seaborn>=0.13.0

# 其他
httpx>=0.25.0
jinja2>=3.1.0
markdown>=3.5.0
tabulate>=0.9.0
tqdm>=4.66.0

# 定时任务与通知
schedule>=1.2.0
